Rename Layouter to LayoutAlgorithm
